requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
PyJWT==2.8.0
python-certifi-win32==1.6.1
pytz==2023.3.post1
PyYAML==6.0.1
secure-smtplib==0.1.1
setuptools-scm==8.0.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
wrapt==1.15.0
